# src/swagger.yaml
openapi: 3.0.0
info:
  title: OKR Express API
  version: 1.0.0
  description: API for managing OKRs, companies, and teams
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        email:
          type: string
        companies:
          type: array
          items:
            type: object
            properties:
              _id:
                type: string
              name:
                type: string
        roles:
          type: array
          items:
            type: object
            properties:
              company:
                type: object
                properties:
                  _id:
                    type: string
                  name:
                    type: string
              role:
                type: string
...
    CheckIn:
      type: object
      properties:
        _id:
          type: string
        okr:
          type: string
        user:
          type: string
        comment:
          type: string
        updates:
          type: array
          items:
            type: object
            properties:
              index:
                type: number
              previousProgress:
                type: number
              newProgress:
                type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
paths:
  /api/auth/register:
    post:
      summary: Register a new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid input
        '409':
          description: Email already exists
  /api/auth/login:
    post:
      summary: Login user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
        '404':
          description: User not found
...
  /api/check-ins:
    post:
      summary: Create a new check-in for an OKR
      tags: [Check-ins]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - okrId
                - updates
              properties:
                okrId:
                  type: string
                updates:
                  type: array
                  items:
                    type: object
                    required:
                      - index
                      - newProgress
                    properties:
                      index:
                        type: number
                        description: Index of the key result to update
                      newProgress:
                        type: number
                        minimum: 0
                        maximum: 100
                        description: New progress value (0-100)
                comment:
                  type: string
                  description: Optional comment about the check-in
      responses:
        '201':
          description: Check-in created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckIn'
        '400':
          description: Invalid key result index or progress value
        '401':
          description: Unauthorized
        '403':
          description: User does not have access to this OKR
        '404':
          description: OKR not found
  /api/check-ins/{okrId}:
    get:
      summary: Get all check-ins for an OKR
      tags: [Check-ins]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: okrId
          required: true
          schema:
            type: string
          description: ID of the OKR
      responses:
        '200':
          description: List of check-ins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CheckIn'
        '401':
          description: Unauthorized
        '403':
          description: User does not have access to this OKR
        '404':
          description: OKR not found
...